sort + uniq
Sorting the list of fonts. And also removing duplicates.

This is VERY useful on my Linux system.

Supposedly, `uniq()` was added on Vim version 7.4.218, which has been a few years ago. I've decided to add that if clause just to improve compatibility just in case someone still runs a very old version.